增加异常处理，倍速发送

+++ src/settingsdialog.h
@@ -86,7 +86,7 @@ public:
         QSerialPort::FlowControl flowControl;
         QString stringFlowControl;
         bool localEchoEnabled;
-        int sendInterval;//报文自动发送间隔
+        float sendInterval;//报文自动发送间隔
     };
 
     explicit SettingsDialog(QWidget *parent = nullptr);
